+++ packages/root/src/types/hickory.ts
@@ -1,4 +1,4 @@
-import { SessionLocation, AnyLocation } from "./location";
+import { SessionLocation, Hrefable } from "./location";
 import { ResponseHandler, ToArgument, NavType } from "./navigate";
 import { LocationUtilOptions } from "./location_utils";
 
@@ -10,7 +10,7 @@ export type HistoryConstructor<O> = (
 
 export interface History {
   location: SessionLocation;
-  to_href(to: AnyLocation): string;
+  href(to: Hrefable): string;
   current(): void;
   cancel(): void;
   destroy(): void;