+++ packages/root/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+- The `stringify` function from `location_utils` can receive a location object or a string.
+- Remove `AnyLocation` type, add `Hrefable` type.
+
 ## 2.0.0-beta.7
 
 - Remove `pathname` option.